+++ internal/boxcli/generate.go
@@ -8,8 +8,8 @@ import (
 	"github.com/spf13/cobra"
 
 	"go.jetpack.io/devbox"
-	"go.jetpack.io/devbox/internal/boxcli/devopt"
 	"go.jetpack.io/devbox/internal/cloud"
+	"go.jetpack.io/devbox/internal/impl/devopt"
 )
 
 type generateCmdFlags struct {
