+++ internal/boxcli/midcobra/telemetry.go
@@ -16,10 +16,10 @@ import (
 	"github.com/spf13/pflag"
 
 	"go.jetpack.io/devbox"
-	"go.jetpack.io/devbox/internal/boxcli/devopt"
 	"go.jetpack.io/devbox/internal/boxcli/featureflag"
 	"go.jetpack.io/devbox/internal/build"
 	"go.jetpack.io/devbox/internal/envir"
+	"go.jetpack.io/devbox/internal/impl/devopt"
 	"go.jetpack.io/devbox/internal/telemetry"
 )
 
@@ -223,9 +223,10 @@ func getPackagesAndCommitHash(c *cobra.Command) ([]string, string) {
 		path = configFlag.Value.String()
 	}
 
-	box, err := devbox.OpenWithoutWarnings(&devopt.Opts{
-		Dir:    path,
-		Writer: os.Stdout,
+	box, err := devbox.Open(&devopt.Opts{
+		Dir:          path,
+		Writer:       os.Stdout,
+		ShowWarnings: false,
 	})
 	if err != nil {
 		return []string{}, ""
