+++ internal/boxcli/services.go
@@ -7,7 +7,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"go.jetpack.io/devbox"
-	"go.jetpack.io/devbox/internal/boxcli/devopt"
+	"go.jetpack.io/devbox/internal/impl/devopt"
 )
 
 type servicesCmdFlags struct {
