+++ internal/boxcli/update.go
@@ -8,7 +8,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"go.jetpack.io/devbox"
-	"go.jetpack.io/devbox/internal/boxcli/devopt"
+	"go.jetpack.io/devbox/internal/impl/devopt"
 )
 
 type updateCmdFlags struct {
